refactor: refactor orderReducer in reducer folder
