+++ src/contexts/orderContext.js
@@ -1,6 +1,7 @@
 import { createContext, useContext, useEffect, useReducer } from "react";
 import axios from "axios";
 import { useAuth } from "./authContext";
+import { orderReducer } from "../reducer/orderReducer";
 
 const OrderContext = createContext();
 
@@ -8,34 +9,6 @@ const initialState = {
     orders: [],
     isLoading: false,
     error: "",
-}
-
-const orderReducer = (state, action) => {
-    switch (action.type) {
-        case "INITIALIZE":
-            return {
-                ...state,
-                isLoading: true,
-                error: ""
-            };
-
-        case "SET_ORDERS":
-            return {
-                ...state,
-                isLoading: false,
-                orders: action.payload
-            };
-
-        case "SET_ERROR":
-            return {
-                ...state,
-                isLoading: false,
-                error: action.payload
-            };
-
-        default:
-            return state;
-    }
 }
 
 const OrderProvider = ({ children }) => {
